+++ src/collective/documentgenerator/helper/base.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import copy
 import datetime
 from collective.documentgenerator.interfaces import IDisplayProxyObject
 from collective.documentgenerator.interfaces import IDocumentGenerationHelper
@@ -32,11 +33,11 @@ class DocumentGenerationHelperView(object):
         proxy_obj.helper_view = self
         return proxy_obj
 
-    def get_value(self, field_name, default=None, as_utf8=False):
+    def get_value(self, field_name, default=None, as_utf8=False):  # pragma: no cover
         """See IDocumentGenerationHelper. To implements."""
         raise NotImplementedError()
 
-    def display(self, field_name, no_value=''):
+    def display(self, field_name, no_value=''):  # pragma: no cover
         """See IDocumentGenerationHelper. To implements."""
         raise NotImplementedError()
 
@@ -59,7 +60,7 @@ class DocumentGenerationHelperView(object):
 
         return safe_unicode(formatted_date)
 
-    def display_voc(self, field_name, separator=','):
+    def display_voc(self, field_name, separator=','):  # pragma: no cover
         """See IDocumentGenerationHelper. To implements."""
         raise NotImplementedError()
 
@@ -82,7 +83,7 @@ class DocumentGenerationHelperView(object):
         display = self.appy_renderer.renderXhtml(html_text)
         return display
 
-    def display_widget(self, fieldname, clean=True, soup=False):
+    def display_widget(self, fieldname, clean=True, soup=False):  # pragma: no cover
         """See IDocumentGenerationHelper. To implements."""
         raise NotImplementedError()
 
@@ -94,7 +95,7 @@ class DocumentGenerationHelperView(object):
     def list(self, field_name):
         return self.get_value(field_name)
 
-    def list_voc(self, field_name, list_keys=False, list_values=True):
+    def list_voc(self, field_name, list_keys=False, list_values=True):  # pragma: no cover
         """See IDocumentGenerationHelper. To implements."""
         raise NotImplementedError()
 
@@ -134,6 +135,30 @@ class DocumentGenerationHelperView(object):
             return ctx[name]
         else:
             return default
+
+    def mailing_list(self):  # pragma: no cover
+        """
+            Return a mailing list (that will be added to generation context in "loop" view) to loop on.
+            Improvement using from plone.memoize import view ?
+        """
+        return []
+
+    def mailed_context(self, mailed_data):
+        """ Modify current context to remove mailing_list and add mailed_data """
+        try:
+            new_context = copy.copy(self.appy_renderer.contentParser.env.context)
+            for key in ('mailing_list', 'mailed_doc'):
+                if key in new_context:
+                    del new_context[key]
+        except:
+            return {'mailed_data': mailed_data}
+        new_context['mailed_data'] = mailed_data
+        return new_context
+
+    def do_mailing(self):
+        """ Check if mailing data must be replaced in template """
+        ctx = self.appy_renderer.contentParser.env.context
+        return 'mailed_data' in ctx
 
 
 class DisplayProxyObject(object):